requirements.txt
uvicorn
fastapi
psycopg2-binary
SQLAlchemy
pydantic
fastapi-utils
passlib[bcrypt]
pyjwt
sqlalchemy-utils
typing_inspect
python-multipart
